refactored utils package on swagger-codegenn module.
