+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/v3/utils/ImplementationVersion.java
@@ -1,4 +1,4 @@
-package io.swagger.codegen.utils;
+package io.swagger.codegen.v3.utils;
 
 public class ImplementationVersion {
     public static String read() {
