+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/php/PhpModelTest.java
@@ -26,7 +26,7 @@ import java.util.Map;
 
 import static io.swagger.codegen.v3.CodegenConstants.IS_ENUM_EXT_NAME;
 import static io.swagger.codegen.languages.helpers.ExtensionHelper.getBooleanValue;
-import static io.swagger.codegen.utils.ModelUtils.updateCodegenPropertyEnum;
+import static io.swagger.codegen.v3.utils.ModelUtils.updateCodegenPropertyEnum;
 
 @SuppressWarnings("static-method")
 public class PhpModelTest {
